Move state transition into state files and generalize chain interface
