+++ go.mod
@@ -3,9 +3,7 @@ module github.com/phoreproject/synapse
 require (
 	github.com/dgraph-io/badger v1.5.5-0.20190226225317-8115aed38f8f
 	github.com/go-test/deep v1.0.1
-	github.com/gogo/protobuf v1.2.1
 	github.com/golang/protobuf v1.3.0
-	github.com/ipfs/go-ipfs-addr v0.0.1
 	github.com/labstack/echo v3.3.10+incompatible
 	github.com/labstack/gommon v0.2.8 // indirect
 	github.com/libp2p/go-libp2p v0.0.1
@@ -18,13 +16,15 @@ require (
 	github.com/libp2p/go-libp2p-pubsub v0.0.1
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/multiformats/go-multiaddr v0.0.1
-	github.com/multiformats/go-multiaddr-net v0.0.1
 	github.com/phoreproject/bls v0.0.0-20190219000203-afaefda3ea64
 	github.com/phoreproject/prysm v0.0.0-20190209202311-c85edb6a4546
 	github.com/prysmaticlabs/prysm v0.0.0-20190306201309-34f0241f3247 // indirect
 	github.com/sirupsen/logrus v1.3.0
 	github.com/valyala/fasttemplate v1.0.0 // indirect
-	golang.org/x/arch v0.0.0-20190312162104-788fe5ffcd8c // indirect
-	golang.org/x/net v0.0.0-20190301231341-16b79f2e4e95
+	golang.org/x/crypto v0.0.0-20190426145343-a29dc8fdc734 // indirect
+	golang.org/x/net v0.0.0-20190424112056-4829fb13d2c6
+	golang.org/x/sync v0.0.0-20190423024810-112230192c58 // indirect
+	golang.org/x/sys v0.0.0-20190429190828-d89cdac9e872 // indirect
+	golang.org/x/text v0.3.2 // indirect
 	google.golang.org/grpc v1.19.0
 )
